minds: create backup/tunnel parity for /api/v1 + repoint UI flows
Extract the create-orchestration helpers (backup-request builder,
post-creation tunnel/account callback, region resolve/persist) from
app.py into a new shared workspace_create.py that both app.py and
api_v1.py import (api_v1 can't import app.py -- that's the cycle). The
agent-facing POST /api/v1/workspaces now accepts the backup_* fields,
provisions restic backups, injects the Cloudflare tunnel token, and
persists the region -- matching the desktop create form.

Repoint the browser UI's lifecycle (start/stop), destroy (settings +
detail-page status poll / SSE log tail / retry), and backup-export flows
onto the versioned /api/v1/workspaces routes. The legacy /api/* JSON
routes stay registered during the transition; the create page
(creating.js) and the backup-status badge keep their existing routes
(the v1 surface intentionally omits the UI-only redirect_url/status_text
and a live backup-status summary).
