+++ apps/minds/changelog/mngr-continue-minds-api.md
@@ -1,0 +1,5 @@
+Wire the agent-facing `/api/v1/workspaces` create route up to full backup + tunnel parity, and point the browser UI's lifecycle / destroy / backup-export flows at the versioned API.
+
+- Extracted the create-orchestration helpers (backup-request builder, post-creation tunnel/account callback, region resolve/persist) out of `app.py` into a new shared `workspace_create.py` so both the browser create form and the `/api/v1/workspaces` create route build the exact same thing. `POST /api/v1/workspaces` now accepts the `backup_*` fields, provisions restic backups, injects the Cloudflare tunnel token, and persists the chosen region just like the desktop form does.
+
+- The landing page's start/stop buttons now call `POST /api/v1/workspaces/<id>/start|stop`, the workspace-settings destroy button and the destroy detail page (status poll + SSE log tail + retry) now use `POST /api/v1/workspaces/<id>/destroy` and `/api/v1/workspaces/operations/<id>`, and backup download now lists snapshots via `GET /api/v1/workspaces/<id>/backups` and exports the most recent via `POST .../backups/<snapshot>/export`. No user-visible behavior change. The legacy `/api/*` JSON routes remain registered during the transition.
