+++ apps/minds/imbue/minds/desktop_client/static/destroying.js
@@ -1,7 +1,10 @@
-// Destroy detail page: polls /api/destroying/<id>/{status,log} every 1s,
-// appends new log content, transitions the badge on terminal status.
-// Reads the agent id from #destroying-page data-agent-id so the template
-// stays JS-free.
+// Destroy detail page: drives the log tail + status badge from the versioned
+// operation resource. Status is polled from
+// /api/v1/workspaces/operations/<id> (authoritative completion signal) and the
+// live log streams over SSE from .../operations/<id>/logs. Retry re-issues the
+// v1 destroy; dismiss removes the on-disk record (no v1 equivalent, so it stays
+// on its dedicated route). Reads the agent id from #destroying-page
+// data-agent-id so the template stays JS-free.
 (function () {
   var pageEl = document.getElementById('destroying-page');
   if (!pageEl) return;
@@ -12,9 +15,9 @@
   var retryBtn = document.getElementById('destroying-retry-btn');
   var dismissBtn = document.getElementById('destroying-dismiss-btn');
 
-  var logOffset = 0;
   var lastStatus = pageEl.getAttribute('data-initial-status') || 'running';
-  var pollTimer = null;
+  var statusPoll = null;
+  var source = null;
   var stopped = false;
 
   function setStatusBadge(status) {
@@ -38,78 +41,86 @@
     logEl.scrollTop = logEl.scrollHeight;
   }
 
-  function fetchLog() {
-    return fetch('/api/destroying/' + agentId + '/log?after=' + logOffset)
-      .then(function (resp) {
-        if (resp.status === 404) return null;
-        return resp.json();
-      })
+  function stopPolling() {
+    if (statusPoll) { clearInterval(statusPoll); statusPoll = null; }
+  }
+
+  function closeSource() {
+    if (source) { source.close(); source = null; }
+  }
+
+  // Apply an authoritative status (lowercased to match the badge vocabulary).
+  // The v1 operation resource reports RUNNING / DONE / FAILED.
+  function applyStatus(status) {
+    if (!status || stopped) return;
+    if (status !== lastStatus) {
+      lastStatus = status;
+      setStatusBadge(status);
+    }
+    if (status === 'done') {
+      stopped = true;
+      stopPolling();
+      closeSource();
+      window.setTimeout(function () { window.location.href = '/'; }, 800);
+    } else if (status === 'failed') {
+      stopped = true;
+      stopPolling();
+      closeSource();
+      actionsEl.classList.remove('hidden');
+    }
+  }
+
+  function pollStatus() {
+    fetch('/api/v1/workspaces/operations/' + encodeURIComponent(agentId))
+      .then(function (resp) { return resp.ok ? resp.json() : null; })
       .then(function (data) {
-        if (!data) return;
-        if (data.content) appendLog(data.content);
-        if (typeof data.next_offset === 'number') logOffset = data.next_offset;
+        if (data && data.status) applyStatus(String(data.status).toLowerCase());
       })
       .catch(function () {});
   }
 
-  function fetchStatus() {
-    return fetch('/api/destroying/' + agentId + '/status')
-      .then(function (resp) {
-        if (resp.status === 404) return null;
-        return resp.json();
-      })
-      .then(function (data) {
-        if (!data) return null;
-        return data.status;
-      })
-      .catch(function () { return null; });
+  // Live log tail. The SSE replays the log from the start on (re)connect and
+  // emits a final {"done": true, "status": ...} frame, which is a secondary
+  // completion signal alongside the status poll.
+  function openSource() {
+    closeSource();
+    source = new EventSource('/api/v1/workspaces/operations/' + encodeURIComponent(agentId) + '/logs');
+    source.onmessage = function (event) {
+      var data;
+      try { data = JSON.parse(event.data); } catch (e) { return; }
+      if (data.log) appendLog(data.log);
+      if (data.done) {
+        closeSource();
+        if (data.status) applyStatus(String(data.status).toLowerCase());
+      }
+    };
+    source.onerror = function () { closeSource(); };
   }
 
-  function tick() {
-    if (stopped) return;
-    Promise.all([fetchLog(), fetchStatus()]).then(function (results) {
-      var status = results[1];
-      if (status && status !== lastStatus) {
-        lastStatus = status;
-        setStatusBadge(status);
-      }
-      if (status === 'done') {
-        stopped = true;
-        // One last log read in case the wrapper printed final lines.
-        fetchLog().then(function () {
-          window.setTimeout(function () { window.location.href = '/'; }, 800);
-        });
-        return;
-      }
-      if (status === 'failed') {
-        stopped = true;
-        actionsEl.classList.remove('hidden');
-        // Pull final log content too.
-        fetchLog();
-        return;
-      }
-      pollTimer = window.setTimeout(tick, 1000);
-    });
+  function start() {
+    stopped = false;
+    setStatusBadge(lastStatus);
+    openSource();
+    pollStatus();
+    statusPoll = setInterval(pollStatus, 1000);
   }
 
   if (retryBtn) {
     retryBtn.addEventListener('click', function () {
       retryBtn.disabled = true;
-      fetch('/api/destroy-agent/' + agentId, { method: 'POST' })
+      fetch('/api/v1/workspaces/' + encodeURIComponent(agentId) + '/destroy', { method: 'POST' })
         .then(function (resp) {
           if (!resp.ok) {
             retryBtn.disabled = false;
             alert('Could not start retry');
             return;
           }
-          // Reset state and start polling again.
+          // Reset state and start the log tail + status poll again.
           logEl.textContent = '';
-          logOffset = 0;
           lastStatus = 'running';
-          stopped = false;
           actionsEl.classList.add('hidden');
-          setStatusBadge('running');
-          tick();
+          retryBtn.disabled = false;
+          start();
         })
         .catch(function () {
           retryBtn.disabled = false;
@@ -121,10 +132,10 @@
   if (dismissBtn) {
     dismissBtn.addEventListener('click', function () {
       dismissBtn.disabled = true;
-      fetch('/api/destroying/' + agentId + '/dismiss', { method: 'POST' })
+      fetch('/api/destroying/' + encodeURIComponent(agentId) + '/dismiss', { method: 'POST' })
         .finally(function () { window.location.href = '/'; });
     });
   }
 
-  tick();
+  start();
 })();
