+++ apps/minds/imbue/minds/desktop_client/templates.py
@@ -1137,11 +1137,11 @@ def render_destroying_page(
 ) -> str:
     """Render the detail page for an in-flight or recently-completed destroy.
 
-    The page polls ``/api/destroying/<agent_id>/{status,log}`` to keep its
-    log tail and status badge up to date; once status flips to ``done`` it
-    redirects to ``/``. ``status`` is the initial server-side computed
-    value (``running``/``failed``/``done``) so the page renders correctly
-    even before the first poll completes.
+    The page polls ``/api/v1/workspaces/operations/<agent_id>`` for status and
+    streams its log over SSE from ``.../operations/<agent_id>/logs``; once
+    status flips to ``done`` it redirects to ``/``. ``status`` is the initial
+    server-side computed value (``running``/``failed``/``done``) so the page
+    renders correctly even before the first poll completes.
     """
     return CATALOG.render(
         "pages.Destroying",
